# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shopify Conectos API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/shopify/products": {
            "post": {
                "tags": [
                    "Shopify Products"
                ],
                "summary": "Sincronizar productos con Shopify",
                "operationId": "1b7de29541cc7d7fb4ffb40327cb013d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "products"
                                ],
                                "properties": {
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "external_id",
                                                "title",
                                                "status",
                                                "variants"
                                            ],
                                            "properties": {
                                                "external_id": {
                                                    "type": "string",
                                                    "example": "PROD-123"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Cámara Hikvision 4K"
                                                },
                                                "vendor": {
                                                    "type": "string",
                                                    "example": "Hikvision"
                                                },
                                                "product_type": {
                                                    "type": "string",
                                                    "example": "Cámaras"
                                                },
                                                "description_html": {
                                                    "type": "string",
                                                    "example": "<strong>Alta definición</strong>"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "active"
                                                },
                                                "tags": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "seguridad",
                                                        "4k"
                                                    ]
                                                },
                                                "images": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "format": "url"
                                                    },
                                                    "example": [
                                                        "https://example.com/img1.jpg"
                                                    ]
                                                },
                                                "variants": {
                                                    "type": "array",
                                                    "items": {
                                                        "required": [
                                                            "sku",
                                                            "price"
                                                        ],
                                                        "properties": {
                                                            "sku": {
                                                                "type": "string",
                                                                "example": "HK-12345"
                                                            },
                                                            "price": {
                                                                "type": "number",
                                                                "example": 1250.5
                                                            },
                                                            "compare_at_price": {
                                                                "type": "number",
                                                                "nullable": true
                                                            },
                                                            "inventory_quantity": {
                                                                "type": "integer",
                                                                "example": 10
                                                            },
                                                            "inventory_management": {
                                                                "type": "string",
                                                                "example": "shopify"
                                                            },
                                                            "requires_shipping": {
                                                                "type": "boolean",
                                                                "example": true
                                                            },
                                                            "taxable": {
                                                                "type": "boolean",
                                                                "example": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "collections": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "example": "gid://shopify/Collection/123456"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Products saved for processing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/shopify/products/uploads/process": {
            "post": {
                "tags": [
                    "Shopify Products"
                ],
                "summary": "Procesar cargas de productos pendientes",
                "operationId": "07fc50631823a91b2cb5121bf4ef465e",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número máximo de productos a procesar",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Process results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/shopify/products/{provider}/{externalId}": {
            "get": {
                "tags": [
                    "Shopify Products"
                ],
                "summary": "Buscar sincronización de producto con categorías",
                "operationId": "ce395f37612d04b9ae49cd89ef6b162e",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "syscom"
                    },
                    {
                        "name": "externalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "12345"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product sync data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/shopify/products/provider/{provider}": {
            "get": {
                "tags": [
                    "Shopify Products"
                ],
                "summary": "Listar productos sincronizados por proveedor",
                "operationId": "7ed34f5e61870f011b7456ce5daf4f1f",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "syscom"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of product syncs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format **Bearer &lt;token&gt;**",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}